Fixed all bugs with the ManualTests: * Assert if IsDBNull is called for a row before getting its value. * TCP timeouts were hard failures, instead of soft failtures that sent timeouts. * Cleaned up SniError. * Assert/crash when recieving a null return value. * Cleaned up DataTestUtility's handling of excepted asserts. * Fixed test bugs and removed tests that were hitting breaking changes. * Enabled manual tests in the powershell script.

+++ src/tests/ManualTests/ProviderAgnostic/ReaderTest/ReaderTest.cs
@@ -15,8 +15,9 @@ namespace System.Data.SqlClient.ManualTesting.Tests
         [CheckConnStrSetupFact]
         public static void TestMain()
         {
-            string connectionString = DataTestUtility.TcpConnStr;
+            string connectionString = new SqlConnectionStringBuilder(DataTestUtility.TcpConnStr) { InitialCatalog = "pubs" }.ToString();
 
+            bool tempTableCreated = false;
             string tempTable = DataTestUtility.GetUniqueName("T", "[", "]");
             string tempKey = DataTestUtility.GetUniqueName("K", "[", "]");
 
@@ -36,6 +37,7 @@ namespace System.Data.SqlClient.ManualTesting.Tests
                         #region <<Create temp table>>
                         cmd.CommandText = "SELECT au_id, au_lname, au_fname, phone, address, city, state, zip, contract into " + tempTable + " from authors where au_id='UNKNOWN-ID'";
                         cmd.ExecuteNonQuery();
+                        tempTableCreated = true;
 
                         cmd.CommandText = "alter table " + tempTable + " add constraint " + tempKey + " primary key (au_id)";
                         cmd.ExecuteNonQuery();
@@ -248,16 +250,19 @@ namespace System.Data.SqlClient.ManualTesting.Tests
             }
             finally
             {
-                using (DbConnection con = provider.CreateConnection())
+                if (tempTableCreated)
                 {
-                    con.ConnectionString = connectionString;
-                    con.Open();
-
-                    using (DbCommand cmd = provider.CreateCommand())
+                    using (DbConnection con = provider.CreateConnection())
                     {
-                        cmd.Connection = con;
-                        cmd.CommandText = "drop table " + tempTable;
-                        cmd.ExecuteNonQuery();
+                        con.ConnectionString = connectionString;
+                        con.Open();
+
+                        using (DbCommand cmd = provider.CreateCommand())
+                        {
+                            cmd.Connection = con;
+                            cmd.CommandText = "drop table " + tempTable;
+                            cmd.ExecuteNonQuery();
+                        }
                     }
                 }
             }
